moving interfaces to component files
