+++ client-src/elements/form-definition.ts
@@ -1,6 +1,116 @@
 import {html, TemplateResult} from 'lit';
 import * as enums from './form-field-enums';
-import {Feature, FormattedFeature, MetadataFields} from './datatypes';
+
+interface BaseBrowser {
+  view: {
+    val: string;
+    text: string;
+    url: string;
+    notes: string;
+  };
+}
+
+interface ChromeBrowser extends BaseBrowser {
+  bug: string;
+  blink_components: string[];
+  devrel: string[];
+  owners: string[];
+  prefixed: boolean;
+  status: {val: string};
+  desktop: string;
+  android: string;
+  webview: string;
+  ios: string;
+}
+
+interface FirefoxBrowser extends BaseBrowser {}
+interface SafariBrowser extends BaseBrowser {}
+interface WebDevBrowser extends BaseBrowser {}
+interface OtherBrowser {
+  view: {notes: string};
+}
+
+interface Resources {
+  docs: string[];
+  samples: string[];
+}
+
+interface Standards {
+  spec: string;
+  maturity: {val: number; text: string};
+}
+
+type FeatureName = string;
+
+interface Feature {
+  name?: FeatureName;
+  category_int: number;
+  enterprise_feature_categories: string[];
+  feature_type_int: number;
+  intent_stage_int: number;
+  standards: Standards;
+  tag_review_status_int: number;
+  security_review_status_int: number;
+  privacy_review_status_int: number;
+  resources: Resources;
+  tags: string[];
+  browsers: {
+    chrome: ChromeBrowser;
+    ff: FirefoxBrowser;
+    safari: SafariBrowser;
+    webdev: WebDevBrowser;
+    other: OtherBrowser;
+  };
+}
+
+export interface FormattedFeature {
+  category: number;
+  enterprise_feature_categories: string[];
+  feature_type: number;
+  intent_stage: number;
+  accurate_as_of: boolean;
+  spec_link: string;
+  standard_maturity: number;
+  tag_review_status: number;
+  security_review_status: number;
+  privacy_review_status: number;
+  sample_links: string[];
+  doc_links: string[];
+  search_tags: string[];
+  blink_components: string;
+  bug_url: string;
+  devrel: string[];
+  owner: string[];
+  prefixed: boolean;
+  impl_status_chrome: string;
+  shipped_milestone: string;
+  shipped_android_milestone: string;
+  shipped_webview_milestone: string;
+  shipped_ios_milestone: string;
+  ff_views: string;
+  ff_views_link: string;
+  ff_views_notes: string;
+  safari_views: string;
+  safari_views_link: string;
+  safari_views_notes: string;
+  web_dev_views: string;
+  web_dev_views_link: string;
+  web_dev_views_notes: string;
+  other_views_notes: string;
+  [key: string]: any; // Allow additional properties
+}
+
+interface Section {
+  name?: string;
+  fields: string[];
+  isImplementationSection?: boolean;
+  implStatusValue?: number | null;
+}
+
+interface MetadataFields {
+  name: string;
+  sections: Section[];
+}
 
 const COMMA_SEPARATED_FIELDS: string[] = [
   'owner',
